requirements.txt
pyreadline3
python-dotenv
google-generativeai
mdv
